FLGO-71 forum subscriptions

+++ app/api/internal/endpoints/show_blog.go
@@ -41,7 +41,7 @@ func (api *API) ShowBlog(r *http.Request) (int, proto.Message) {
 
 	offset := limit * (page - 1)
 
-	dbResponse, err := api.services.DB().FetchBlogTopics(r.Context(), uint32(blogID), uint32(limit), uint32(offset))
+	dbResponse, err := api.services.DB().FetchBlogTopics(r.Context(), blogID, limit, offset)
 
 	if err != nil {
 		if dbtools.IsNotFoundError(err) {
@@ -56,6 +56,6 @@ func (api *API) ShowBlog(r *http.Request) (int, proto.Message) {
 		}
 	}
 
-	blog := datahelpers.GetBlog(dbResponse, uint32(page), uint32(limit), api.config)
+	blog := datahelpers.GetBlog(dbResponse, page, limit, api.config)
 	return http.StatusOK, blog
 }
