+++ app/helpers/appspecific.go
@@ -9,7 +9,7 @@ func IsValidLimit(limit uint64) bool {
 	return limit >= 5 && limit <= 50
 }
 
-func CalculatePageCount(totalCount, limit uint32) uint32 {
+func CalculatePageCount(totalCount, limit uint64) uint64 {
 	if limit < 1 {
 		return 0
 	}
@@ -23,7 +23,7 @@ func CalculatePageCount(totalCount, limit uint32) uint32 {
 	return pageCount
 }
 
-func GetGender(userId uint32, userSex uint8) pb.Common_Gender {
+func GetGender(userId uint64, userSex uint8) pb.Common_Gender {
 	if userId > 0 {
 		if userSex == 0 {
 			return pb.Common_FEMALE
@@ -35,19 +35,19 @@ func GetGender(userId uint32, userSex uint8) pb.Common_Gender {
 	return pb.Common_UNKNOWN_GENDER
 }
 
-func GetUserAvatarUrl(baseURL string, userId uint32, photoNumber uint32) string {
+func GetUserAvatarUrl(baseURL string, userId uint64, photoNumber uint64) string {
 	var avatar string
 
 	if photoNumber != 0 {
-		userId := strconv.FormatUint(uint64(userId), 10)
-		photoNumber := strconv.FormatUint(uint64(photoNumber), 10)
+		userId := strconv.FormatUint(userId, 10)
+		photoNumber := strconv.FormatUint(photoNumber, 10)
 		avatar = baseURL + "/users/" + userId + "_" + photoNumber
 	}
 
 	return avatar
 }
 
-func GetCommunityAvatarUrl(baseURL string, communityId uint32) string {
-	id := strconv.FormatUint(uint64(communityId), 10)
+func GetCommunityAvatarUrl(baseURL string, communityId uint64) string {
+	id := strconv.FormatUint(communityId, 10)
 	return baseURL + "/communities/" + id + ".jpg"
 }
